Only conversations missing

+++ main.py
@@ -1,10 +1,10 @@
 from itakello_logging import ItakelloLogging
 
-from src.managers.action_m import ActionManager
-from src.managers.conversation_m import ConversationManager
-from src.managers.database_m import DatabaseManager
-from src.managers.experiment_m import ExperimentManager
-from src.managers.input_m import InputManager
+from llm_simulator.managers.action_m import ActionManager
+from llm_simulator.managers.conversation_m import ConversationManager
+from llm_simulator.managers.database_m import DatabaseManager
+from llm_simulator.managers.experiment_m import ExperimentManager
+from llm_simulator.managers.input_m import InputManager
 
 ItakelloLogging(
     debug=True,
@@ -30,10 +30,10 @@ def main() -> None:
             break
     db_m.select_database()
 
-    input_m = InputManager()
-    action_m = ActionManager()
-    experiment_m = ExperimentManager(db_m=db_m)  # ⚒️
-    conversation_m = ConversationManager(db_m=db_m)  # ❌
+    input_m = InputManager()  # ✅
+    action_m = ActionManager()  # ✅
+    experiment_m = ExperimentManager(db_m=db_m)  # ✅
+    conversation_m = ConversationManager(db_m=db_m)  # ⚒️
 
     while True:
         action = action_m.select_initial_action()
@@ -50,16 +50,21 @@ def main() -> None:
         while True:
             action = action_m.select_experiment_action()
             if action == "Perform new conversations":  # ⚒️
-                experiment_m.perform_conversations(experiment)
+                conversation_m.perform_conversations(experiment)
                 continue
             elif action == "Duplicate and update experiment":  # ✅
-                experiment = experiment_m.update_experiment(experiment)
+                experiment = experiment_m.duplicate_and_update_experiment(experiment)
                 if experiment != None:
                     logger.info(f"\nUpdated experiment:\n\n{experiment}")
                 break
             elif action == "Select old conversations":  # ✅
                 conversation = conversation_m.select_conversation(experiment)
             elif action == "Delete experiment":  # ✅
+                if experiment.creator != db_m.username:
+                    logger.warning(
+                        "You are not the creator of this experiment. You cannot delete it."
+                    )
+                    continue
                 if input_m.confirm("Are you sure you want to delete this experiment?"):
                     experiment_m.delete_experiment(experiment)
                     break
@@ -75,7 +80,15 @@ def main() -> None:
                     # ui_m.update_conversation(conversation_dict)
                     pass
                 elif action == "Delete conversation":  # ❌
-                    # ui_m.delete_conversation(experiment, conversation_dict)
+                    if conversation.creator != db_m.username:
+                        logger.warning(
+                            "You are not the creator of this conversation. You cannot delete it."
+                        )
+                        continue
+                    if input_m.confirm(
+                        "Are you sure you want to delete this conversation?"
+                    ):
+                        conversation_m.delete_conversation(experiment, conversation)
                     break
                 else:  # Go back
                     break
